fix unit tests

+++ harness/src/test/scala/com/robertjneal/rl/AgentTest.scala
@@ -38,7 +38,7 @@ class AgentTest {
 
     assertEquals(agent.e.possibleStateActions, result1.e.possibleStateActions)
     assertEquals(
-      1,
+      1L,
       result1
         .actionSteps(OneState)
         .values
@@ -46,7 +46,7 @@ class AgentTest {
         .sum - actionRewards.size.toLong
     )
     assertEquals(
-      result1.step,
+      result1.step.toInt.toLong,
       result1
         .actionSteps(OneState)
         .values
@@ -58,7 +58,7 @@ class AgentTest {
 
     assertEquals(result1.e.possibleStateActions, result2.e.possibleStateActions)
     assertEquals(
-      2,
+      2L,
       result2
         .actionSteps(OneState)
         .values
@@ -66,7 +66,7 @@ class AgentTest {
         .sum - actionRewards.size.toLong
     )
     assertEquals(
-      result2.step,
+      result2.step.toInt.toLong,
       result2
         .actionSteps(OneState)
         .values
